Player can now talk to the NPC
The NPCs text lines are stored in a text file in "Scripts" as the format
"NpcName_action". Each time the player talks to the NPC, the next text
line will be displayed onto the screen until the last one is displayed,
then the counter will reset.

The next task is to make the text only display on the screen for a few
seconds (maybe 2-5) and then dissapear on it's own. I had a way of doing
it in my old version but it was very messy.